+++ modules-tools/prov-template-archetype/src/main/resources/archetype-resources/app-template-service/src/test/java/AppTest.java
@@ -1,7 +1,0 @@
-import junit.framework.TestCase;
-public class AppTest extends TestCase {
-    public void testApp() {
-        App.main(null);
-        assertTrue(true);
-    }
-}