+++ modules-tools/prov-template-archetype/src/main/resources/oldstuff/service/dispatch/EnactorException.java
@@ -1,0 +1,9 @@
+#set( $symbol_pound = '#' )
+#set( $symbol_dollar = '$' )
+#set( $symbol_escape = '\' )
+package org.openprovenance.prov.service.dispatch;
+
+public class EnactorException extends RuntimeException {
+    public EnactorException(String s, String method) {
+    }
+}